(patch) Removed faulty construct table code
made the matrix math its own class to seperate concerns.
removed the code that displays the steps as the way that revised work it makes it difficult to keep track of teh variables.
the solution is correct because I implemented rounding, the same solution I provided to normal primal simplex earlier on in development.